requirements.txt
numpy==1.24.3
Pillow==10.1.0
matplotlib==3.8.0
scikit-learn==1.3.2
tqdm==4.66.1
kaggle==1.5.16
python-dotenv==1.1.1
streamlit
